backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.20
alembic>=1.12.0,<1.13
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
redis>=5.0.1
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
email-validator>=2.1.0

